+++ delta_trace_db/db/delta_trace_db_collection.py
@@ -13,7 +13,7 @@ _logger = logging.getLogger(__name__)
 
 class Collection(CloneableFile):
     class_name = "Collection"
-    version = "15"
+    version = "16"
 
     def __init__(self):
         """