+++ delta_trace_db/query/nodes/logical_node.py
@@ -1,11 +1,21 @@
 # coding: utf-8
-from typing import Any, Dict
+from typing import Any, Dict, override
 from delta_trace_db.query.nodes.enum_node_type import EnumNodeType
 from delta_trace_db.query.nodes.query_node import QueryNode
 
 
 class AndNode(QueryNode):
     def __init__(self, conditions: list[QueryNode]):
+        """
+        (en) Query node for AND operation.
+
+        (ja) AND演算のためのクエリノード。
+
+        Parameters
+        ----------
+        conditions: list[QueryNode]
+            A list of child nodes.
+        """
         self.conditions = conditions
 
     @classmethod
@@ -13,9 +23,11 @@ class AndNode(QueryNode):
         from delta_trace_db.query.nodes.query_node import QueryNode
         return cls([QueryNode.from_dict(dict(e)) for e in src['conditions']])
 
+    @override
     def evaluate(self, data: dict) -> bool:
         return all(c.evaluate(data) for c in self.conditions)
 
+    @override
     def to_dict(self) -> dict:
         return {
             'type': EnumNodeType.and_.name,
@@ -26,6 +38,16 @@ class AndNode(QueryNode):
 
 class OrNode(QueryNode):
     def __init__(self, conditions: list[QueryNode]):
+        """
+        (en) Query node for Or operation.
+
+        (ja) Or演算のためのクエリノード。
+
+        Parameters
+        ----------
+        conditions: list[QueryNode]
+            A list of child nodes.
+        """
         self.conditions = conditions
 
     @classmethod
@@ -33,9 +55,11 @@ class OrNode(QueryNode):
         from delta_trace_db.query.nodes.query_node import QueryNode
         return cls([QueryNode.from_dict(dict(e)) for e in src['conditions']])
 
+    @override
     def evaluate(self, data: dict) -> bool:
         return any(c.evaluate(data) for c in self.conditions)
 
+    @override
     def to_dict(self) -> dict:
         return {
             'type': EnumNodeType.or_.name,
@@ -45,7 +69,17 @@ class OrNode(QueryNode):
 
 
 class NotNode(QueryNode):
-    def __init__(self, condition):
+    def __init__(self, condition: QueryNode):
+        """
+        (en) Query node for Not operation.
+
+        (ja) Not演算のためのクエリノード。
+
+        Parameters
+        ----------
+        condition: QueryNode
+            A child node.
+        """
         self.condition = condition
 
     @classmethod
@@ -53,15 +87,18 @@ class NotNode(QueryNode):
         from delta_trace_db.query.nodes.query_node import QueryNode
         return cls(QueryNode.from_dict(dict(src['condition'])))
 
+    @override
     def evaluate(self, data: dict) -> bool:
         return not self.condition.evaluate(data)
 
+    @override
     def to_dict(self) -> dict:
         return {
             'type': EnumNodeType.not_.name,
             'condition': self.condition.to_dict(),
             'version': '1',
         }
+
 
 __all__ = [
     "AndNode",