+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.37
+
+* Added docs.
+* Refactoring was performed.
+* Fixed an issue where the return type hints for initiated_at and finalized_at in TemporalTrace were incorrect.
+* Fixed an issue where the Actor class did not inherit CloneableFile.
+
+
 ## 0.0.36
 
 * Refactoring was performed.